Refactor Admin Bundle event classes - related to #14642
