+++ bundles/AdminBundle/Event/Model/DocumentDeleteInfoEvent.php
@@ -13,8 +13,9 @@
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Event\Model;
+namespace Pimcore\Bundle\AdminBundle\Event\Model;
 
+use Pimcore\Event\Model\DocumentEvent;
 use Pimcore\Event\Traits\ElementDeleteInfoEventTrait;
 
 /**
